# Conflicts:
#	iris/src/main/java/ir/mohaymen/iris/chat/Chat.java
#	iris/src/main/java/ir/mohaymen/iris/chat/ChatController.java
#	iris/src/main/java/ir/mohaymen/iris/chat/ChatRepository.java
#	iris/src/main/java/ir/mohaymen/iris/chat/ChatServiceImpl.java
#	iris/src/main/java/ir/mohaymen/iris/contact/ContactController.java
#	iris/src/main/java/ir/mohaymen/iris/message/Message.java
#	iris/src/main/java/ir/mohaymen/iris/message/MessageController.java
#	iris/src/main/java/ir/mohaymen/iris/message/MessageRepository.java
#	iris/src/main/java/ir/mohaymen/iris/message/MessageServiceImpl.java
#	iris/src/main/java/ir/mohaymen/iris/permission/Permission.java
#	iris/src/main/java/ir/mohaymen/iris/permission/PermissionService.java
#	iris/src/main/java/ir/mohaymen/iris/permission/PermissionServiceImpl.java
#	iris/src/main/java/ir/mohaymen/iris/profile/ProfileController.java
#	iris/src/main/java/ir/mohaymen/iris/search/contact/SearchContactController.java
#	iris/src/main/java/ir/mohaymen/iris/subscription/SubscriptionController.java
#	iris/src/main/java/ir/mohaymen/iris/subscription/SubscriptionRepository.java
#	iris/src/main/java/ir/mohaymen/iris/subscription/SubscriptionService.java
#	iris/src/main/java/ir/mohaymen/iris/subscription/SubscriptionServiceImpl.java

+++ iris/src/main/java/ir/mohaymen/iris/chat/ChatController.java
@@ -3,6 +3,10 @@ package ir.mohaymen.iris.chat;
 import ir.mohaymen.iris.contact.ContactService;
 import ir.mohaymen.iris.message.Message;
 import ir.mohaymen.iris.message.MessageService;
+import ir.mohaymen.iris.profile.ProfileDto;
+import ir.mohaymen.iris.profile.ProfileMapper;
+import ir.mohaymen.iris.permission.Permission;
+import ir.mohaymen.iris.permission.PermissionService;
 import ir.mohaymen.iris.profile.*;
 import ir.mohaymen.iris.subscription.Subscription;
 import ir.mohaymen.iris.subscription.SubscriptionService;
@@ -10,6 +14,7 @@ import ir.mohaymen.iris.user.User;
 import ir.mohaymen.iris.user.UserService;
 import ir.mohaymen.iris.utility.BaseController;
 import ir.mohaymen.iris.utility.Nameable;
+import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -22,9 +27,8 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.text.MessageFormat;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
+
 @RestController
 @RequestMapping("/api/chats")
 @RequiredArgsConstructor
@@ -36,14 +40,46 @@ public class ChatController extends BaseController {
     private final ModelMapper modelMapper;
     private final ChatProfileService chatProfileService;
     private final MessageService messageService;
+    private final PermissionService permissionService;
     private final ContactService contactService;
     private final UserProfileService userProfileService;
     private final Logger logger = LoggerFactory.getLogger(ChatController.class);
 
+    @PutMapping("/edit-chat-info")
+    public ResponseEntity<?> editChatInfo(@RequestBody EditChatDto editChatDto) {
+        Chat chat = chatService.getById(editChatDto.getId());
+        chat.setBio(editChatDto.getBio());
+        chat.setTitle(editChatDto.getTitle());
+        chat.setLink(editChatDto.getLink());
+        if (chat.isPublic() && !editChatDto.isPublic())
+            chat.setLink(UUID.randomUUID().toString());
+        chat.setPublic(editChatDto.isPublic());
+        chatService.createOrUpdate(chat);
+        return ResponseEntity.ok().build();
+    }
+
     @PostMapping("/create-chat")
+    @Transactional
     public ResponseEntity<GetChatDto> createChat(@RequestBody @Valid CreateChatDto createChatDto) {
+        if (createChatDto.getChatType() == ChatType.PV) {
+            if (createChatDto.getUserIds().size() != 1) {
+                if (createChatDto.getUserIds().contains(getUserByToken().getUserId())){
+                    logger.error("PV chat can't have owner in adding list");
+                    throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+                }
+                logger.error("PV chat must have one user id in user ids");
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            }
+            Chat chat;
+            if ((chat = subscriptionService.PVExistance(getUserByToken().getUserId() , createChatDto.getUserIds().stream().toList().get(0) )) != null) {
+                logger.info("Found a already existing PV chat. returning the chat.");
+                return getGetChatDtoResponseEntity(chat);
+            }
+        }
         Chat chat = modelMapper.map(createChatDto, Chat.class);
-        logger.info(MessageFormat.format("user with phone number:{0} wants to create chat ", getUserByToken().getPhoneNumber()));
+
+        logger.info(MessageFormat.format("user with phone number:{0} wants to create chat ",
+                getUserByToken().getPhoneNumber()));
         if (!chat.getChatType().equals(ChatType.PV) && chat.getTitle().isBlank()) {
             logger.error("non-PV chat must have title");
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
@@ -51,53 +87,55 @@ public class ChatController extends BaseController {
         chat.setCreatedAt(Instant.now());
         if (chat.getChatType() == ChatType.PV)
             chat.setPublic(false);
-        createInternalSub(chat , getUserByToken());
         chat = chatService.createOrUpdate(chat);
-        if ((createChatDto.getUserIds().size() != 1 && chat.getChatType() == ChatType.PV))
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        for (Long id : createChatDto.getUserIds()) {
-            User user = userService.getById(id);
-            if (chatService.isInChat(chat, user))
-                throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        Set<Permission> ownerPermissions = chat.getChatType() == ChatType.PV
+                ? Permission.getDefaultPermissions(chat.getChatType())
+                : Permission.getOwnerPermissions();
+        createInternalSub(chat, getUserByToken(), ownerPermissions);
+        for (Long userId : createChatDto.getUserIds()) {
             try {
-                createInternalSub(chat, user);
+                createInternalSub(chat, User.builder().userId(userId).build());
             } catch (Exception ex) {
                 throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
             }
         }
+        logger.info("chat is created");
         return getGetChatDtoResponseEntity(chat);
     }
 
     @GetMapping("/get-chat/{id}")
     public ResponseEntity<GetChatDto> getChat(@PathVariable Long id) {
         Chat chat = chatService.getById(id);
-        if (!chatService.isInChat(chat, getUserByToken()) || (!chat.isPublic() && chat.getChatType() != ChatType.PV))
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        if (!chat.isPublic() && !chatService.isInChat(chat, getUserByToken()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         return getGetChatDtoResponseEntity(chat);
     }
 
     @GetMapping("/{link}")
     public ResponseEntity<GetChatDto> getChatByLink(@PathVariable String link) {
-        if (link == null)
+        if (link == null || link.isBlank())
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         Chat chat = chatService.getByLink(link);
-        if (!chatService.isInChat(chat, getUserByToken()) || (!chat.isPublic() && chat.getChatType() != ChatType.PV))
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        if (!chat.isPublic() && !chatService.isInChat(chat, getUserByToken()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         return getGetChatDtoResponseEntity(chat);
     }
 
     private ResponseEntity<GetChatDto> getGetChatDtoResponseEntity(Chat chat) {
+        logger.info(MessageFormat.format("chat with id:{0} is fetched by user:{1}", chat.getChatId(),
+                getUserByToken().getPhoneNumber()));
         GetChatDto getChatDto = modelMapper.map(chat, GetChatDto.class);
         List<ProfileDto> profileDtoList = new ArrayList<>();
         if (chat.getChatType() != ChatType.PV) {
             List<ChatProfile> chatProfileList = chatProfileService.getByChat(chat);
-            if (chatProfileList != null)
+            if (!chatProfileList.isEmpty())
                 chatProfileList.forEach(chatProfile -> profileDtoList.add(ProfileMapper.mapToProfileDto(chatProfile)));
         } else {
             User user = userService.getById(chatService.getOtherPVUser(chat, getUserByToken().getUserId()));
             List<UserProfile> userProfileList = userProfileService.getByUser(user);
-            if (userProfileList != null)
-               userProfileList.forEach(chatProfile -> profileDtoList.add(ProfileMapper.mapToProfileDto(chatProfile)));
+            if (userProfileList != null || !userProfileList.isEmpty())
+                user.getProfiles()
+                        .forEach(chatProfile -> profileDtoList.add(ProfileMapper.mapToProfileDto(chatProfile)));
         }
         getChatDto.setChatId(chat.getChatId());
         getChatDto.setProfileDtoList(profileDtoList);
@@ -109,21 +147,21 @@ public class ChatController extends BaseController {
         Chat chat = sub.getChat();
         MenuChatDto menuChatDto = modelMapper.map(chat, MenuChatDto.class);
         if (chat.getChatType() != ChatType.PV) {
-            List<ChatProfile> chatProfileList = chatProfileService.getByChat(chat);
-            if (chatProfileList.size() != 0)
-                menuChatDto.setMedia(chatProfileList.get(chatProfileList.size() - 1).getMedia());
+            ChatProfile chatProfile = chatProfileService.getLastChatProfile(chat);
+            if (chatProfile != null )
+                menuChatDto.setMedia(chatProfile.getMedia());
         }
         else {
             User user = userService.getById(chatService.getOtherPVUser(chat, getUserByToken().getUserId()));
             Nameable nameable = subscriptionService.setName(contactService.getContactByFirstUser(getUserByToken()), user);
-            List<UserProfile> userProfileList = userProfileService.getByUser(user);
             menuChatDto.setTitle(nameable.fullName());
-            if (userProfileList.size() != 0)
-                menuChatDto.setMedia(userProfileList.get(userProfileList.size() - 1).getMedia());
+            UserProfile userProfile = userProfileService.getLastUserProfile(user);
+            if (userProfile != null)
+                menuChatDto.setMedia(userProfile.getMedia());
             menuChatDto.setUserFirstName(nameable.getFirstName());
         }
-        if (chat.getMessages().size() != 0) {
-            Message message = messageService.getLastMessageByChatId(chat);
+        Message message = messageService.getLastMessageByChatId(chat);
+        if (message != null) {
             menuChatDto.setUnSeenMessages(messageService.countUnSeenMessages(sub.getLastMessageSeenId(), chat.getChatId()));
             menuChatDto.setLastMessage(message.getText());
             menuChatDto.setSentAt(message.getSendAt());
@@ -133,8 +171,7 @@ public class ChatController extends BaseController {
                 menuChatDto.setMedia(user.getProfiles().get(user.getProfiles().size() - 1).getMedia());
             }
             menuChatDto.setUserFirstName(nameable.getFirstName());
-        }
-        else {
+        } else {
             menuChatDto.setSentAt(chat.getCreatedAt());
         }
         return menuChatDto;
@@ -146,7 +183,7 @@ public class ChatController extends BaseController {
         for (Subscription sub : subscriptionService.getAllSubscriptionByUserId(getUserByToken().getUserId())) {
             menuChatDtos.add(createMenuChat(sub));
         }
-        Collections.sort(menuChatDtos, (o1, o2) -> {
+        menuChatDtos.sort((o1, o2) -> {
             if (o1.getSentAt().isBefore(o2.getSentAt()))
                 return 1;
             return -1;
@@ -156,21 +193,23 @@ public class ChatController extends BaseController {
 
     @DeleteMapping("delete-chat/{id}")
     public ResponseEntity<?> deleteChat(@PathVariable Long id) throws Exception {
-        if (!chatService.isInChat(chatService.getById(id), getUserByToken()))
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        if (!chatService.isInChat(chatService.getById(id), getUserByToken())
+                || !permissionService.hasAccess(getUserByToken().getUserId(), id, Permission.OWNER))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         chatService.deleteById(id);
         return new ResponseEntity<>(null, HttpStatus.OK);
     }
 
     private Subscription createInternalSub(Chat chat, User user) {
+        return createInternalSub(chat, user, Permission.getDefaultPermissions(chat.getChatType()));
+    }
+
+    private Subscription createInternalSub(Chat chat, User user, Set<Permission> permissions) {
         Subscription sub = new Subscription();
         sub.setChat(chat);
         sub.setUser(user);
-        subscriptionService.createOrUpdate(sub);
-        return sub;
+        sub.setPermissions(permissions);
+        return subscriptionService.createOrUpdate(sub);
     }
 
 }
-
-
-
